requirements.txt
click
grayskull
jinja2
PyGithub
PyYAML
requests
